requirements.txt
pyobjc-core>=10.0
pyobjc-framework-Cocoa>=10.0
pyyaml>=6.0
speechrecognition>=3.10.0
pyaudio>=0.2.11

